Move RigWellResultFrame implementation into separate file
Update/correct includes accordingly
