+++ ApplicationLibCode/ModelVisualization/RivWellSpheresPartMgr.cpp
@@ -22,6 +22,7 @@
 
 #include "RigMainGrid.h"
 #include "RigSimWellData.h"
+#include "RigWellResultFrame.h"
 #include "RigWellResultPoint.h"
 
 #include "RimEclipseCase.h"
