+++ ApplicationLibCode/ProjectDataModel/Flow/RimFlowDiagSolution.cpp
@@ -24,6 +24,7 @@
 #include "RigFlowDiagResults.h"
 #include "RigMainGrid.h"
 #include "RigSimWellData.h"
+#include "RigWellResultFrame.h"
 #include "RigWellResultPoint.h"
 
 #include "RimEclipseResultCase.h"
