+++ ApplicationLibCode/ProjectDataModel/RimEclipseView.cpp
@@ -38,6 +38,7 @@
 #include "RigResultAccessorFactory.h"
 #include "RigSimWellData.h"
 #include "RigVirtualPerforationTransmissibilities.h"
+#include "RigWellResultFrame.h"
 #include "RigWellResultPoint.h"
 
 #include "Rim2dIntersectionView.h"
