+++ ApplicationLibCode/ProjectDataModel/Streamlines/RimStreamlineInViewCollection.cpp
@@ -28,7 +28,7 @@
 #include "RigResultAccessorFactory.h"
 #include "RigSimWellData.h"
 #include "RigTracerPoint.h"
-#include "RigWellResultPoint.h"
+#include "RigWellResultFrame.h"
 
 #include "RimEclipseCase.h"
 #include "RimEclipseInputCase.h"
