+++ ApplicationLibCode/ReservoirDataModel/CMakeLists_files.cmake
@@ -90,6 +90,7 @@ set(SOURCE_GROUP_HEADER_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigPressureDepthData.h
     ${CMAKE_CURRENT_LIST_DIR}/RigMswCenterLineCalculator.h
     ${CMAKE_CURRENT_LIST_DIR}/RigWellAllocationOverTime.h
+    ${CMAKE_CURRENT_LIST_DIR}/RigWellResultFrame.h
 )
 
 set(SOURCE_GROUP_SOURCE_FILES
@@ -177,6 +178,7 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigPressureDepthData.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigMswCenterLineCalculator.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigWellAllocationOverTime.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/RigWellResultFrame.cpp
 )
 
 list(APPEND CODE_HEADER_FILES ${SOURCE_GROUP_HEADER_FILES})
