+++ ApplicationLibCode/ReservoirDataModel/RigSimulationWellCenterLineCalculator.cpp
@@ -26,7 +26,7 @@
 #include "RigEclipseCaseData.h"
 #include "RigMainGrid.h"
 #include "RigSimWellData.h"
-#include "RigWellResultPoint.h"
+#include "RigWellResultFrame.h"
 
 #include "RimEclipseCase.h"
 #include "RimEclipseView.h"
